requirements.txt
# Core dependencies
requests==2.31.0
numpy==1.24.3
scipy==1.11.1
faiss-cpu==1.7.4
sentence-transformers==2.2.2
openai>=1.6.1,<2.0.0
flask==2.3.2
flask-cors==4.0.0

# Optional dependencies for enhanced functionality
langchain>=0.1.0
langchain-openai>=0.0.2
nltk==3.8.1
spacy==3.6.1

# Development and testing
pytest==7.4.0
pytest-cov==4.1.0
black==23.7.0
flake8==6.0.0

# Logging and utilities
loguru==0.7.0
python-dotenv==1.0.0
click==8.1.6

# Data processing
pandas==2.0.3
scikit-learn==1.3.0

# Web framework
gunicorn==21.2.0
waitress==2.1.2